# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MEETable API Docs",
    "description": "MEETable api 문서입니다",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "name": "authorization",
      "type": "apiKey",
      "in": "header"
    }
  },
...
  "paths": {
...
    "/member/info": {
      "get": {
        "tags": [
          "member"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/member/resetname": {
      "patch": {
        "tags": [
          "member"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "seungho111"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/member/quit": {
      "delete": {
        "tags": [
          "member"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }, "404": {
            "description": "Not Found"
          }, "500": {
            "description": "백엔드 에러"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  }
}